+++ src/penalties.cpp
@@ -22,6 +22,7 @@
 #include <vector>
 #include <math.h>
 
+#include "recon_structs.h"
 #include "penalties.h"
 
 #define SQRT2 1.41421356237309504880168872420969807
@@ -33,7 +34,7 @@ inline int sign(double x){
         return 1;
 }
 
-double func(double du, double alpha, double beta, double delta, double lambda, double * recon_volume,
+double func(double du, double alpha, double beta, double delta, double lambda, float * recon_volume,
             double *wgts, std::size_t*neighbor_indices, int q, const int num_neighbors,size_t array_size);
 
 void initialize_2d_weights(struct iterative_params *ip){
@@ -58,7 +59,7 @@ void initialize_2d_weights(struct iterative_params *ip){
     
 }
 
-double quadratic(int curr_idx,struct iterative_params * ip, double *recon_volume){
+double quadratic(int curr_idx,struct iterative_params * ip, float *recon_volume){
     // Compute pixel update using 2D (In-plane) QUADRATIC PENALTY function
     
     // Identify indices of voxels we'll compute 
@@ -85,7 +86,7 @@ double quadratic(int curr_idx,struct iterative_params * ip, double *recon_volume
 }
 
 
-double edge_preserving(int curr_idx,struct iterative_params * ip, double * recon_volume){
+double edge_preserving(int curr_idx,struct iterative_params * ip, float * recon_volume){
    // Identify indices of voxels we'll compute 
     std::size_t neighbor_indices[8];
     
@@ -138,7 +139,7 @@ double edge_preserving(int curr_idx,struct iterative_params * ip, double * recon
     }    
 }
 
-double func(double du, double alpha, double beta, double delta, double lambda, double * recon_volume,
+double func(double du, double alpha, double beta, double delta, double lambda, float * recon_volume,
             double *wgts, std::size_t *neighbor_indices, int q, int num_neighbors,size_t array_size){
 	double sum = 0;
 	for (int i = 0; i < num_neighbors; i++){
